requirements.txt
numpy
optuna
gymnasium
stable-baselines3
pandas
matplotlib
torch
tensorboardX
torchvision
Pillow
